LPS-70123 Consistency with Bundle-SymbolicName
